COMP: Signed-unsigned comparison in ExtractMetricImageSlice.

+++ Applications/RegisterImageToTubesUsingRigidTransform/ExtractMetricImageSlice.cxx
@@ -68,7 +68,7 @@ int main( int argc, char * argv[] )
   MetricImageType::IndexType extractionIndex =
     extractionRegion.GetIndex();
   unsigned int indexCount = 0;
-  for( int ii = 0; ii < MetricDimension; ++ii )
+  for( int ii = 0; ii < static_cast< int >( MetricDimension ); ++ii )
     {
     if( ii != sliceDirections[0] && ii != sliceDirections[1] )
       {
